+++ global-client/src/components/ui/HubIcon.vue
@@ -8,8 +8,8 @@
 		</div>
 		<Icon v-if="pinnable" type="plus" class="text-green absolute right-0" @click.prevent="pin"></Icon>
 		<Icon v-if="pinned" type="remove" class="text-red absolute right-0 hidden group-hover:block" @click.prevent="remove"></Icon>
-		<Icon :type="type" :size="size" class="text-white mx-auto"></Icon>
-		<HubLogo v-if="hub" :hub-url="hub.url" :hub-id="hub.hubId" :change-to-dark="false" class="absolute h-14 w-14 top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2"></HubLogo>
+		<!-- <Icon :type="type" :size="size" class="text-white mx-auto"></Icon> -->
+		<HubLogo v-if="hub" :hub-url="hub.url" :hub-id="hub.hubId" :change-to-dark="false" class="h-20 w-20 mx-auto"></HubLogo>
 	</div>
 </template>