+++ global-client/src/pages/App.vue
@@ -10,8 +10,8 @@
 					<Modal :show="global.isModalVisible">
 						<div class="flex flex-col justify-between h-full">
 							<div class="flex-1 text-center">
-								<router-link to="/" v-slot="{ isActive }">
-									<HubIcon type="pubhubs-home" :active="isActive" class="text-white" @click="toggleMenu.toggleMenu()"></HubIcon>
+								<router-link to="/">
+									<Icon type="pubhubs-home" size="3xl" class="text-white mx-auto" @click="toggleMenu.toggleMenu()"></Icon>
 								</router-link>
 
 								<Line v-if="global.loggedIn && global.hasPinnedHubs" class="m-6 mt-8"></Line>