+++ hub-client/src/components/rooms/RoomEvent.vue
@@ -7,7 +7,7 @@
 		<!-- Normal Event -->
 		<div v-else class="group flex gap-4 pl-6 pr-3 transition-all duration-150 ease-in-out hover:bg-lightgray-light hover:dark:bg-hub-background-2 py-4">
 			<Avatar :userId="event.sender"></Avatar>
-			<div class="w-4/5 md:w-3/5">
+			<div class="w-4/5 xl:w-3/5">
 				<div class="flex flex-wrap items-center">
 					<div class="relative flex flex-wrap items-center w-full gap-x-2 md:w-fit pr-2 min-h-6">
 						<UserDisplayName :user="event.sender"></UserDisplayName>