+++ hub-client/src/components/rooms/RoomTimeline.vue
@@ -261,6 +261,7 @@
 			allMessagesLoaded = await pubhubs.loadOlderEvents(props.room);
 			numLoadedMessages = props.room.timelineGetNumMessageEvents();
 		}
+		await scrollToLastReadEvent();
 	}
 </script>